RSA: Move `verification::Key` to `public::Key`.
Take a step towards having RSA keys support encryption/decryption
in addition to signing/verification.

No functional changes are intended.

Some documentation was moved around.
